requirements.txt
Flask
redis
celery
Werkzeug>=2.0
Jinja2==2.11.1
gunicorn==20.0.4
eventlet==0.30.2
flask-cors
Flask-SocketIO==5.1.1
flasgger
Flask-SQLAlchemy
importlib_resources
Flask-WTF
flower
pymysql
depinfo
pipdeptree
python-dotenv
python-engineio==4.2.1
python-socketio==5.4.0
future
contrabass
